Fixed update of items for which auto/burst does not apply when swapped into the hand Moved helicopter in Estoni so that it is unblocked by a weed and can therefore appear completely Fixed conditions where the game should remain paused but doesn't, and where it should unpause but doesn't
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@1531 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Strategic/Campaign Types.h
@@ -4,9 +4,9 @@
 #include "types.h"
 
 //Macro to convert sector coordinates (1-16,1-16) to 0-255
-#define SECTOR(x,y)			((y-1)*16+x-1)
-#define SECTORX(SectorID)	((SectorID % 16) + 1)
-#define SECTORY(SectorID)	((SectorID / 16) + 1)
+#define SECTOR(x,y)			(UINT8)((y-1)*16+x-1)
+#define SECTORX(SectorID)	(UINT8)((SectorID % 16) + 1)
+#define SECTORY(SectorID)	(UINT8)((SectorID / 16) + 1)
 
 //Sector enumerations
 //
